+++ R/zzz.r
@@ -23,6 +23,7 @@
    assign("col.fg",          "#7a6d59",        envir=.chesstrainer)
    assign("col.square.l",    "#f0d9b5",        envir=.chesstrainer)
    assign("col.square.d",    "#b58863",        envir=.chesstrainer)
+   assign("col.square.be",   "#6b6b6b",        envir=.chesstrainer)
    assign("col.top",         "#b09d7f",        envir=.chesstrainer)
    assign("col.bot",         "#b09d7f",        envir=.chesstrainer)
    assign("col.help",        "#b09d7f",        envir=.chesstrainer)
@@ -37,6 +38,33 @@
    # set default font
 
    assign("font.mono", "mono", envir=.chesstrainer)
+
+   # create starting position matrix 'pos'
+
+   pos <- matrix("", 8, 8)
+   pos[2,] <- "WP"
+   pos[1,1] <- pos[1,8] <- "WR"
+   pos[1,2] <- pos[1,7] <- "WN"
+   pos[1,3] <- pos[1,6] <- "WB"
+   pos[1,4] <- "WQ"
+   pos[1,5] <- "WK"
+   pos[7,] <- "BP"
+   pos[8,1] <- pos[8,8] <- "BR"
+   pos[8,2] <- pos[8,7] <- "BN"
+   pos[8,3] <- pos[8,6] <- "BB"
+   pos[8,4] <- "BQ"
+   pos[8,5] <- "BK"
+   colnames(pos) <- LETTERS[1:8]
+   rownames(pos) <- 1:8
+   assign("pos", pos, envir=.chesstrainer)
+
+   pos <- rbind("", pos, "")
+   pos <- cbind("", pos, "")
+   rownames(pos) <- 0:9
+   colnames(pos) <- 0:9
+   pos[1,3:8]  <- c("WK","WQ","WR","WB","WN","WP")
+   pos[10,3:8] <- c("BK","BQ","BR","BB","BN","BP")
+   assign("boardeditorpos", pos, envir=.chesstrainer)
 
    # welcome message